requirements.txt
python-pptx>=0.6.22
pyinstaller==6.3.0
tkinterdnd2==0.3.0
PyPDF2>=3.0.0
Pillow>=10.0.0 